frontend/requirements.txt
certifi==2024.7.4
chardet==4.0.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
mysqlclient==2.2.4
requests==2.25.1
typing_extensions==4.12.2
urllib3==1.26.19
